requirements.txt
Flask==2.3.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0
tiktoken==0.5.1
scikit-learn==1.3.0
torch==2.0.1
transformers==4.33.2

pygame==2.5.2
pydub==0.25.1
SpeechRecognition==3.10.0

ffmpeg-python==0.2.0

Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

python-dotenv==1.0.0
requests==2.31.0
